% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_jcr_forest}
\alias{build_jcr_forest}
\title{Build the joint classification/regression forest}
\usage{
build_jcr_forest(x, maxnodes)
}
\arguments{
\item{x}{A single integer.}
}
\description{
Build the joint classification/regression forest
}

