% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_generation.R
\name{data_gen}
\alias{data_gen}
\title{Create dummy data x}
\usage{
data_gen(N, Nx = 3)
}
\arguments{
\item{N}{number of observations}

\item{Nx}{number of explanatory variables}
}
\description{
Function used for simulating data for test runs
}

