\name{JCRForest-package}
\alias{JCRForest-package}
\alias{JCRForest}
\docType{package}
\title{
\packageTitle{JCRForest}
}
\description{
\packageDescription{JCRForest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{JCRForest}
\packageIndices{JCRForest}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{JCRForest}

Maintainer: \packageMaintainer{JCRForest}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
