% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JCRForest.R
\name{mixed_forest}
\alias{mixed_forest}
\title{Main function}
\usage{
mixed_forest(x, y, mtry, Ntree)
}
\arguments{
\item{x}{Data frame containing explanatory variables}

\item{y}{Data frame containing 2 response variables, one being discrete (factor), the other continuous}

\item{mtry}{Number of features examined per tree}

\item{Ntree}{Number of trees in the forest}
}
\description{
Function used for fitting mixed random forests.
}

